chore: refactor button to separate file
